mandatory finished

+++ headers/utils.h
@@ -6,7 +6,7 @@
 /*   By: ahanaf <ahanaf@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 05:59:28 by ahanaf            #+#    #+#             */
-/*   Updated: 2024/05/26 11:48:03 by ahanaf           ###   ########.fr       */
+/*   Updated: 2024/06/02 15:22:31 by ahanaf           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *str, int fd);
 size_t	ft_strlen(const char *s);
 char	*ft_itoa(unsigned long n);
-void	ft_sleep(int time);
+void	ft_sleep(unsigned long time);
 size_t	get_time(void);
 
 #endif
